refactor(mobile): extract tab screens into src/screens
Move each (tabs)/*.tsx screen body into src/screens/<id>-screen.tsx with
a named export; replace each route file with a thin re-export. No behavior
change.
